requirements.txt
google-generativeai>=0.8.0
opencv-python>=4.6.0
numpy>=1.21.0
Pillow>=9.0.0
PyMuPDF>=1.23.0
reportlab>=3.6.0
flask>=2.2.0
flask-cors>=3.0.0
python-dotenv>=0.19.0
werkzeug>=2.2.0
pypdf>=5.0.0
pywin32>=306
docx2pdf>=0.1.8
python-docx>=0.8.11
docxtpl>=0.16.7
psutil>=5.9.0 